% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sc_artwork.R
\name{get_sc_artwork}
\alias{get_sc_artwork}
\title{Get SoundCloud track urls}
\usage{
get_sc_artwork(track_url = NULL, ci = NULL, size = "t300x300.jpg")
}
\arguments{
\item{track_url}{character. sound track urls.}

\item{ci}{a character. vector that SoundCloud API client id.}

\item{size}{image file size.}
}
\description{
SoundCloud API wrapper function.
}
\examples{
\dontrun{
ci <- '<your client id>'
track_url <- c('http://ift.tt/1Vjezde')
track_url \%>\% purrr::flatmap(get_sc_artwork, ci)
}
}

