% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtheme_Publication.R
\name{theme_Publication}
\alias{theme_Publication}
\title{ggplot2 custom theme Koundinya Desiraju's Publication'}
\usage{
theme_Publication(base_size = 14, base_family = "Helvetica", ...)
}
\references{
https://github.com/koundy/ggplot_theme_Publication
}

